#29: Move segmented control to top.
The repo had a great image in the `App Architecture Overview` section. That image showed us that LocationWeatherView is the only view.

In that file, I looked for the init and then went to configure subviews, and finally to the vStack declaration. There, it was simple enough to move the segmentControl from the end of the `arrangedSubviews` array to the beginning.