#29: Keep using guards.
This is a great example of how simplifying code will unearth errors. If URLSession wasn't returning a valid HTTPURLResponse (for whatever reason), the previous code was not handling it _at all_.

We are not here to fix all issues, so just add a warning and keep going with our task.